refactor(config): jwt, oauth 설정을 AppProperties로 통합
- 기존 JwtProperties 로직을 AppProperties로 변경
- OAuth 설정 추가
