--- src/main/java/dev/wgrgwg/somniverse/security/jwt/properties/JwtProperties.java
@@ -1,30 +1,0 @@
-package dev.wgrgwg.somniverse.security.jwt.properties;
-
-import io.jsonwebtoken.Jwts;
-import jakarta.annotation.PostConstruct;
-import java.nio.charset.StandardCharsets;
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@ConfigurationProperties(prefix = "spring.jwt")
-@Getter
-@Setter
-public class JwtProperties {
-
-    private String secret;
-    private Long accessTokenExpirationMs;
-    private Long refreshTokenExpirationMs;
-
-    private SecretKey secretKey;
-
-    @PostConstruct
-    private void init() {
-        this.secretKey = new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8),
-            Jwts.SIG.HS256.key().build().getAlgorithm());
-    }
-}
